requirements.txt
alabaster==0.7.16
annotated-types==0.6.0
anyio==4.6.2.post1
apturl==0.5.2
asgiref==3.8.1
asttokens==2.4.1
astunparse==1.6.3
attrs==23.2.0
Babel==2.14.0
bcrypt==3.2.0
beautifulsoup4==4.12.3
bleach==6.1.0
blinker==1.8.2
Brlapi==0.8.3
cachetools==5.5.1
certifi==2020.6.20
chardet==4.0.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.4
comm==0.2.2
contourpy==1.2.0
cryptography==3.4.8
cupshelpers==1.0
cycler==0.12.1
dbus-python==1.2.18
debugpy==1.8.1
decorator==5.1.1
defer==1.0.6
defusedxml==0.7.1
dill==0.3.8
distlib==0.3.8
distro==1.7.0
distro-info==1.1+ubuntu0.2
Django==5.1.3
django-cors-headers==4.6.0
django-environ==0.11.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
docutils==0.20.1
duplicity==0.8.21
exceptiongroup==1.2.0
executing==2.0.1
fasteners==0.14.1
fastjsonschema==2.19.1
fica==0.3.1
filelock==3.15.4
Flask==3.0.3
fonttools==4.50.0
future==0.18.2
google-ai-generativelanguage==0.6.15
google-api-core==2.24.1
google-api-python-client==2.160.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.4
googleapis-common-protos==1.66.0
grpcio==1.70.0
grpcio-status==1.70.0
h11==0.14.0
httpcore==1.0.6
httplib2==0.20.2
httpx==0.27.2
idna==3.3
imagesize==1.4.1
importlib-metadata==4.6.4
ipykernel==6.29.3
ipylab==1.0.0
ipython==8.22.2
ipywidgets==8.1.2
itsdangerous==2.2.0
jedi==0.19.1
jeepney==0.7.1
Jinja2==3.1.3
jiter==0.7.0
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter_client==8.6.1
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
jupyterlab_widgets==3.0.10
jupytext==1.16.1
keyring==23.5.0
kiwisolver==1.4.5
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lockfile==0.12.2
louis==3.20.0
macaroonbakery==1.3.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.3
matplotlib-inline==0.1.6
mdit-py-plugins==0.4.0
mdurl==0.1.2
mistune==3.0.2
monotonic==1.6
more-itertools==8.10.0
nbclient==0.10.0
nbconvert==7.16.2
nbformat==5.10.3
nest-asyncio==1.6.0
numpy==1.26.4
oauthlib==3.2.0
olefile==0.46
openai==1.54.3
otter-grader==5.5.0
packaging==24.0
pandas==2.2.1
pandocfilters==1.5.1
paramiko==2.9.3
parso==0.8.3
pexpect==4.8.0
Pillow==9.0.1
platformdirs==4.2.0
pluginbase==1.0.1
prompt-toolkit==3.0.43
proto-plus==1.26.0
protobuf==5.29.3
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycairo==1.20.1
pycups==2.0.1
pydantic==2.6.4
pydantic_core==2.16.3
Pygments==2.17.2
PyGObject==3.42.1
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyparsing==2.4.7
pyRFC3339==1.1
python-apt==2.4.0+ubuntu4
python-dateutil==2.9.0.post0
python-debian==0.1.43+ubuntu1.1
python-decouple==3.8
python-dotenv==1.0.1
python-on-whales==0.70.0
pytz==2022.1
pyxdg==0.27
PyYAML==5.4.1
pyzmq==25.1.2
referencing==0.34.0
reportlab==3.6.8
requests==2.31.0
rpds-py==0.18.0
rsa==4.9
seaborn==0.13.2
SecretStorage==3.3.1
six==1.16.0
sniffio==1.3.1
snowballstemmer==2.2.0
soupsieve==2.5
SpeechRecognition==3.10.1
Sphinx==7.2.6
sphinxcontrib-applehelp==1.0.8
sphinxcontrib-devhelp==1.0.6
sphinxcontrib-htmlhelp==2.0.5
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.7
sphinxcontrib-serializinghtml==1.1.10
sqlparse==0.5.1
stack-data==0.6.3
systemd-python==234
tinycss2==1.2.1
toml==0.10.2
tornado==6.4
tqdm==4.66.2
traitlets==5.14.2
typer==0.9.0
typing_extensions==4.12.2
tzdata==2024.1
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
unattended-upgrades==0.1
uritemplate==4.1.1
urllib3==1.26.5
usb-creator==0.3.7
vboxapi==1.0
virtualenv==20.26.3
wadllib==1.3.6
wafw00f==2.1.0
wcwidth==0.2.13
webencodings==0.5.1
Werkzeug==3.0.4
widgetsnbextension==4.0.10
wrapt==1.16.0
xdg==5
xkit==0.0.0
zipp==1.0.0
